Changes to avoid warnings in windows builds... 1. Avoided a (very unlikely) /0 error risk (when states for all     the sites in a sequence are unknown)     in IQTree::generateProximityTree. 2. Casts to avoid signed/unsigned conversion warnings, in     Windows builds, in:     (a) PlacementRun::doneBatch     (b) IQTree::generateProximityTree     (c) ParsimonyRouter::selectSample     (d) ParsimonyRouter::bestSampleBranch     (e) ParsimonyRouter::updateBranch     (f) ParsimonyRouter::insertCandidate     (g) ParsimonyRouter::constructTree     (h) PhyloTree::getMemoryRequired     (i) PhyloTree::addNewTaxaToTree 3. Casts to avoid double/integer conversion warnings, in     Windows builds, in:     (a) ParsimonyMatrix::calculateLeafParsimonyDistances     (b) PhyloTree::doParsimonySearch     (c)  PhyloTree::addNewTaxaToTree 4. Changed parameters, to avoid conversion warnings, in      Windows builds, in:      (a) ParsimonyMove::consider 5. Changed vector types (to avoid conversion warnings)     (a) PhyloTree::computeParsimonyTreeBatch

+++ tree/parsimonynni.cpp
@@ -45,7 +45,7 @@ public:
         s << " exchanging neighbours of branch " << source_branch_id;
         return s.str();
     }
-    void consider(int branch_id, PhyloNode* leftNode,
+    void consider(intptr_t branch_id, PhyloNode* leftNode,
                   const PhyloBranch& middleBranch,
                   PhyloNode* rightNode, double move_benefit) {
         if (benefit < move_benefit) {
